sticker face/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
insightface
Pillow
